+++ app/src/main/java/app/visto/AppInfo.kt
@@ -3,6 +3,6 @@ package app.visto
 object AppInfo {
     const val APP_NAME = "Visto"
     const val PACKAGE_NAME = "app.visto"
-    const val VERSION_NAME = "1.1.16"
-    const val VERSION_CODE = 37
+    const val VERSION_NAME = "1.1.17"
+    const val VERSION_CODE = 38
 }